Past Tabs and Current Tabs and closing tabs working properly

+++ app/src/main/java/com/cs407/tapntab/BillItem.kt
@@ -1,3 +1,3 @@
 package com.cs407.tapntab
 
-data class BillItem(val name: String, val price: Double, var qty: Int)
+data class BillItem(val name: String, val price: Double, var qty: Long)
