+++ app/src/main/java/com/cs407/tapntab/TotalBillItem.kt
@@ -1,3 +1,3 @@
 package com.cs407.tapntab
 
-data class TotalBillItem(val name: String, val price: Double, var qty: Int)
+data class TotalBillItem(val name: String, val price: Double, var qty: Long)
